Set the margin according to the alignment selected (#928)
As reported in #927, the style of the profile class only considers a
left-margin regardless of the alignment option selected.
Thus, when the user changes the default alignment there is no
corresponding margin to match their selection.
My commit addresses this by enabling the corresponding margin for the
left or right alignment options [ci skip]